pushing to verify working bits
