+++ index.html
@@ -8,7 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>PokeGoDex</title>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.8.0/css/bulma.min.css">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="mj_style.css">
     <script defer src="https://use.fontawesome.com/releases/v5.3.1/js/all.js"></script>
     <script src="https://code.jquery.com/jquery.js"></script>
 
@@ -119,7 +119,7 @@
     </section>
 
     <script src="https://code.jquery.com/jquery.js"></script>
-    <script src="script.js"></script>
+    <script src="pl_script.js"></script>
 
 </body>
